requirements.txt
tensorflow==1.14.0
psycopg2-binary==2.7.5
Pillow==5.2.0
numpy==1.15.4
requests==2.20.1
tqdm==4.23.0
hydrosdk==2.0.0rc9
hydro-serving-grpc==2.0.3
hs==2.1.0rc6
namesgenerator==0.3
kfp==0.1.24
scikit-learn==0.20.2
mlflow==1.0.0
boto3~=1.9.139
sagemaker==1.34.1
jsonschema==2.6.0
docker==3.7.0
paramiko>=2.4.2